Create diagnostics package
